chore: remove RpcDriver type
